requirements.txt
PyQt5>=5.15.0
cryptography>=41.0.0
pandas>=2.0.0
openpyxl>=3.1.0
python-docx>=1.1.0
pdfplumber>=0.10.0
pytest>=8.0.0
